Add practice creation form to homepage - fixes #8
